stage-2/requirements.txt
fastapi
uvicorn[standard]
azure-ai-projects==1.0.0b4
azure-identity
